Modelo final

+++ user_router.py
@@ -1,30 +1,19 @@
 from fastapi import FastAPI, HTTPException,APIRouter, Depends
 from pydantic import BaseModel, EmailStr
 from datetime import timedelta, datetime, timezone
-from models import usuario
 from dependecis import sessao, verificar_token
 from config import bcrypt_context, ALGORITHM, ACESS_TOKEN_EXPIREM,SECRET_KEY
 from jose import jwt, JWSError
 from fastapi.security import OAuth2PasswordRequestForm
 from verifications import verificar_excluir, verificar_usuario
+from datetime import datetime, timedelta, timezone
+from jose import jwt
+from models import UsuarioEntrada
+from database import Usuario
 
 user_router=APIRouter(tags=["user"])
 
-class UsuarioEntrada(BaseModel):
-    nome: str
-    email: EmailStr
-    senha: str
-    admin: bool
-    foto: str | None = None
-    bio: str  | None = None
-
-from datetime import datetime, timedelta, timezone
-from jose import jwt
-
-def criar_token(
-    id_usuario,
-    data_exp=timedelta(minutes=int(ACESS_TOKEN_EXPIREM))
-):
+def criar_token(id_usuario, data_exp=timedelta(minutes=int(ACESS_TOKEN_EXPIREM))):
     data_expiracao = datetime.now(timezone.utc) + data_exp
 
     dict_info = {
@@ -42,27 +31,24 @@ def criar_token(
 
 @user_router.get("/usarios", status_code=200)
 async def listarusuarios(session= Depends(sessao)):
-    Usuario= session.query(usuario).all()
-    return Usuario
+    usuarios= session.query(Usuario).all()
+    return usuarios
 
 @user_router.get("/usuario", status_code=200)
 async def buscarusuario(id_usuario: int, session= Depends(sessao)):
-    user= session.query(usuario).filter(usuario.id == id_usuario).first()
+    user= session.query(Usuario).filter(Usuario.id == id_usuario).first()
     return user
 
 @user_router.post("/usuario", status_code=201)
 async def cadastro(dados: UsuarioEntrada, session= Depends(sessao)):
     
-    Email= session.query(usuario).filter(usuario.email == dados.email).first()
+    Email= session.query(Usuario).filter(Usuario.email == dados.email).first()
 
     if Email:
         raise HTTPException(409, "email já cadastrado!")
-    print(repr(dados.senha))
-    print(type(dados.senha))
-    print(len(dados.senha))
     dados.senha= bcrypt_context.hash(dados.senha)
     
-    u= usuario(**dados.model_dump())
+    u= Usuario(**dados.model_dump())
 
     session.add(u)
     session.commit()
@@ -73,7 +59,7 @@ async def cadastro(dados: UsuarioEntrada, session= Depends(sessao)):
 
 @user_router.post("/login", status_code=200)
 async def login(dadoos_formulario:OAuth2PasswordRequestForm = Depends() , session= Depends(sessao)):
-    user= session.query(usuario).filter(usuario.email == dadoos_formulario.username).first()
+    user= session.query(Usuario).filter(Usuario.email == dadoos_formulario.username).first()
     
     if not user:
 
@@ -83,19 +69,51 @@ async def login(dadoos_formulario:OAuth2PasswordRequestForm = Depends() , sessio
 
         raise HTTPException(status_code=404, detail=("Senha invalida"))
     
-    acess_token= criar_token(user.id)
+    access_token= criar_token(user.id,)
     refresh_token= criar_token(user.id, data_exp= timedelta(days=7))
 
     session.refresh(user)
     return {
-        "access_token": acess_token,
-        "refresh token": refresh_token,
-        "token_type": "bearer"
+        "access_token": access_token,
+        "token_type": "bearer",
+        "refresh token": refresh_token
+    }
+
+@user_router.post("/login-refresh")
+async def login_refresh(dados_formulario: OAuth2PasswordRequestForm = Depends(), session=Depends(sessao)):
+
+    user = session.query(Usuario).filter(
+        Usuario.email == dados_formulario.username
+    ).first()
+
+    if not user:
+        raise HTTPException(404, "Usuário não encontrado")
+
+    if not bcrypt_context.verify(
+        dados_formulario.password,
+        user.senha
+    ):
+        raise HTTPException(401, "Senha inválida")
+
+    access_token = criar_token(
+        user.id,
+        tipo="access"
+    )
+
+    refresh_token = criar_token(
+        user.id,
+        tipo="refresh",
+        data_exp=timedelta(days=7)
+    )
+
+    return {
+        "access_token": access_token,
+        "refresh_token": refresh_token
     }
 
 
 @user_router.patch("/usuario/atualizarperfil", status_code=200)
-async def editar_perfil(bio:str, foto:str, session= Depends(sessao), user: usuario = Depends(verificar_token)):
+async def editar_perfil(bio:str, foto:str, session= Depends(sessao), user: Usuario = Depends(verificar_token)):
     
     user.bio=bio
     user.foto= foto
@@ -106,9 +124,13 @@ async def editar_perfil(bio:str, foto:str, session= Depends(sessao), user: usuar
     return user
 
 @user_router.delete("/usuario/{id_usuario}", status_code=200)
-async def removerusuario(id_usuario:int, user: usuario = Depends(verificar_token),  session=Depends(sessao)):
+async def removerusuario(id_usuario:int, user: Usuario = Depends(verificar_token),  session=Depends(sessao)):
 
-    u= session.query(usuario).filter(usuario.id == id_usuario).first()
+    u= session.query(Usuario).filter(Usuario.id == id_usuario).first()
+
+    if not u:
+        raise HTTPException(status_code=400, detail="usuario não existe")
+    
     verificar_excluir(u.id, user, session)
 
     session.delete(u)
@@ -116,10 +138,10 @@ async def removerusuario(id_usuario:int, user: usuario = Depends(verificar_token
 
     return u
 
-@user_router.get("/refreseh")
-async def use_refresh_token(user: usuario = Depends(verificar_token)):
+@user_router.get("/refresh")
+async def use_refresh_token(token):
 
-    acess_token= criar_token(user.id)
+    acess_token = criar_token(token)
 
     return {
             "token de acesso": acess_token,
